+++ backend/epic_app/admin.py
@@ -5,7 +5,13 @@ from django.contrib import admin
 from django.shortcuts import redirect, render
 from django.urls import path
 
-from epic_app.importers import EpicAgencyImporter, EpicDomainImporter, EpicImporter
+from epic_app.importers import (
+    BaseEpicImporter,
+    EpicAgencyImporter,
+    EpicDomainImporter,
+    EvolutionQuestionImporter,
+    NationalFrameworkQuestionImporter,
+)
 from epic_app.models.epic_questions import (
     Answer,
     EvolutionQuestion,
@@ -73,7 +79,7 @@ class ImportEntityAdmin(admin.ModelAdmin):
         return render(request, "admin/csv_form.html", payload)
 
     @abc.abstractmethod
-    def get_importer(self) -> EpicImporter:
+    def get_importer(self) -> BaseEpicImporter:
         raise NotImplementedError("Should be implemented in concrete class.")
 
 
@@ -95,13 +101,66 @@ class AgencyAdmin(ImportEntityAdmin):
         return EpicAgencyImporter()
 
 
+class NfqAdmin(ImportEntityAdmin):
+    """
+    National Framework Question admin page to allow CSV import.
+    """
+
+    def get_importer(self) -> NationalFrameworkQuestionImporter:
+        return NationalFrameworkQuestionImporter()
+
+
+class EvoAdmin(ImportEntityAdmin):
+    """
+    Evolution Question admin page to allow CSV import.
+    """
+
+    def get_importer(self) -> EvolutionQuestionImporter:
+        return EvolutionQuestionImporter()
+
+
+class LnkAdmin(admin.ModelAdmin):
+    # Admin pages.
+    change_list_template = "generate_changelist.html"
+
+    def get_urls(self):
+        """
+        Extends the default get_urls so we can inject the import-csv on
+
+        Returns:
+            List[str]: A list of the urls to load from the admin page.
+        """
+        urls = super().get_urls()
+        my_urls = [
+            path("generate/", self.generate_links),
+        ]
+        return my_urls + urls
+
+    def generate_links(self, request):
+        """
+        Imports a csv file into the EPIC database structure.
+
+        Args:
+            request (HTTPRequest): HTML request.
+
+        Returns:
+            HTTPRequest: HTML response.
+        """
+        LinkagesQuestion.objects.all().delete()
+        LinkagesQuestion.generate_linkages()
+        self.message_user(
+            request, "Generated one linkage question per existent program"
+        )
+        return redirect("..")
+
+
 # Models exposed to the admin page .
 admin.site.register(EpicUser)
 admin.site.register(Area, AreaAdmin)
 admin.site.register(Agency, AgencyAdmin)
 admin.site.register(Group)
 admin.site.register(Program)
-admin.site.register(NationalFrameworkQuestion)
-admin.site.register(EvolutionQuestion)
-admin.site.register(LinkagesQuestion)
+admin.site.register(NationalFrameworkQuestion, NfqAdmin)
+admin.site.register(EvolutionQuestion, EvoAdmin)
+admin.site.register(LinkagesQuestion, LnkAdmin)
 admin.site.register(Answer)