+++ backend/epic_app/tests/importers/test_domain_csv_importer.py
@@ -1,22 +1,24 @@
 from io import BytesIO
-from pathlib import Path
 
 import pytest
 from django.core.files.uploadedfile import InMemoryUploadedFile
 
-from epic_app.importers import EpicAgencyImporter, EpicDomainImporter, EpicImporter
-from epic_app.models.models import Agency, Area, Group, Program
+from epic_app.importers import BaseEpicImporter, EpicDomainImporter
+from epic_app.importers.csv_base_importer import ProtocoEpicImporter
+from epic_app.models.models import Area, Group, Program
+from epic_app.tests import test_data_dir
 
 
+@pytest.mark.django_db
 class TestEpicDomainImporter:
     def test_epic_domain_importer(self):
-        importer = EpicDomainImporter()
-        assert isinstance(importer, EpicImporter)
+        domain_importer = EpicDomainImporter()
+        assert isinstance(domain_importer, BaseEpicImporter)
+        assert isinstance(domain_importer, ProtocoEpicImporter)
 
-    @pytest.mark.django_db
     def test_import_csv_from_filepath(self):
         # Define test data
-        test_file = Path(__file__).parent / "test_data" / "initial_epic_data.csv"
+        test_file = test_data_dir / "initial_epic_data.csv"
         assert test_file.is_file()
 
         # Verify initial expectations
@@ -36,7 +38,7 @@ class TestEpicDomainImporter:
         # Verify final expectations
         assert len(Area.objects.all()) == 5
         assert len(Group.objects.all()) == 11
-        assert len(Program.objects.all()) == 43
+        assert len(Program.objects.all()) == 42
         assert any(
             [p.description != "" for p in Program.objects.all()]
         ), "No descriptions were imported."
@@ -45,10 +47,9 @@ class TestEpicDomainImporter:
         assert dummy_group not in Group.objects.all()
         assert dummy_program not in Program.objects.all()
 
-    @pytest.mark.django_db
     def test_import_csv_from_inmemoryuploaddedfile(self):
         # Define request.
-        test_file = Path(__file__).parent / "test_data" / "initial_epic_data.csv"
+        test_file = test_data_dir / "initial_epic_data.csv"
         assert test_file.is_file()
         with test_file.open("rb") as csv_file:
             file_io = BytesIO(csv_file.read())
@@ -80,7 +81,7 @@ class TestEpicDomainImporter:
         # Verify final expectations
         assert len(Area.objects.all()) == 5
         assert len(Group.objects.all()) == 11
-        assert len(Program.objects.all()) == 43
+        assert len(Program.objects.all()) == 42
         assert any(
             [p.description != "" for p in Program.objects.all()]
         ), "No descriptions were imported."
@@ -89,41 +90,3 @@ class TestEpicDomainImporter:
         assert dummy_area not in Area.objects.all()
         assert dummy_group not in Group.objects.all()
         assert dummy_program not in Program.objects.all()
-
-
-class TestEpicAgencyImporter:
-    @pytest.fixture(autouse=True)
-    @pytest.mark.django_db
-    def default_epic_domain_data(self):
-        """
-        Fixture to load the predefined database so we can test importing agencies correctly.
-        """
-        # Define test data
-        test_file = Path(__file__).parent / "test_data" / "initial_epic_data.csv"
-        assert test_file.is_file()
-        EpicDomainImporter().import_csv(test_file)
-
-    @pytest.mark.django_db
-    def test_epic_agency_importer(self):
-        agency_importer = EpicAgencyImporter()
-        assert isinstance(agency_importer, EpicImporter)
-
-    @pytest.mark.django_db
-    def test_import_csv_from_filepath(self):
-        # Define test data
-        test_file = Path(__file__).parent / "test_data" / "agency_data.csv"
-        assert test_file.is_file()
-
-        # Verify initial expectations
-        dummy_agency = Agency(name="dummyAgency")
-        dummy_agency.save()
-        assert len(Agency.objects.all()) == 1
-
-        # Run test
-        EpicAgencyImporter().import_csv(test_file)
-
-        # Verify final expectations
-        assert len(Agency.objects.all()) == 6
-
-        # Verify the initial data has been removed.
-        assert dummy_agency not in Agency.objects.all()